add EthTransactionArgs in NewTxArgs interface
