+++ api/api_ethereum.go
@@ -17,7 +17,6 @@
 package api
 
 import (
-	"bytes"
 	"context"
 	"encoding/binary"
 	"errors"
@@ -35,7 +34,6 @@ import (
 	"github.com/klaytn/klaytn/blockchain/vm"
 	"github.com/klaytn/klaytn/common"
 	"github.com/klaytn/klaytn/common/hexutil"
-	"github.com/klaytn/klaytn/common/math"
 	"github.com/klaytn/klaytn/crypto"
 	"github.com/klaytn/klaytn/governance"
 	"github.com/klaytn/klaytn/networks/rpc"
@@ -1031,282 +1029,6 @@ func newEthTransactionReceipt(header *types.Header, tx *types.Transaction, b Bac
 	return fields, nil
 }
 
-// EthTransactionArgs represents the arguments to construct a new transaction
-// or a message call.
-// TransactionArgs in go-ethereum has been renamed to EthTransactionArgs.
-// TransactionArgs is defined in go-ethereum's internal package, so TransactionArgs is redefined here as EthTransactionArgs.
-type EthTransactionArgs struct {
-	From                 *common.Address `json:"from"`
-	To                   *common.Address `json:"to"`
-	Gas                  *hexutil.Uint64 `json:"gas"`
-	GasPrice             *hexutil.Big    `json:"gasPrice"`
-	MaxFeePerGas         *hexutil.Big    `json:"maxFeePerGas"`
-	MaxPriorityFeePerGas *hexutil.Big    `json:"maxPriorityFeePerGas"`
-	Value                *hexutil.Big    `json:"value"`
-	Nonce                *hexutil.Uint64 `json:"nonce"`
-
-	// We accept "data" and "input" for backwards-compatibility reasons.
-	// "input" is the newer name and should be preferred by clients.
-	// Issue detail: https://github.com/ethereum/go-ethereum/issues/15628
-	Data  *hexutil.Bytes `json:"data"`
-	Input *hexutil.Bytes `json:"input"`
-
-	// Introduced by AccessListTxType transaction.
-	AccessList *types.AccessList `json:"accessList,omitempty"`
-	ChainID    *hexutil.Big      `json:"chainId,omitempty"`
-}
-
-// from retrieves the transaction sender address.
-func (args *EthTransactionArgs) from() common.Address {
-	if args.From == nil {
-		return common.Address{}
-	}
-	return *args.From
-}
-
-// data retrieves the transaction calldata. Input field is preferred.
-func (args *EthTransactionArgs) data() []byte {
-	if args.Input != nil {
-		return *args.Input
-	}
-	if args.Data != nil {
-		return *args.Data
-	}
-	return nil
-}
-
-// setDefaults fills in default values for unspecified tx fields.
-func (args *EthTransactionArgs) setDefaults(ctx context.Context, b Backend) error {
-	if args.GasPrice != nil && (args.MaxFeePerGas != nil || args.MaxPriorityFeePerGas != nil) {
-		return errors.New("both gasPrice and (maxFeePerGas or maxPriorityFeePerGas) specified")
-	}
-	// After london, default to 1559 uncles gasPrice is set
-	head := b.CurrentBlock().Header()
-	isMagma := head.BaseFee != nil
-
-	fixedBaseFee := new(big.Int).SetUint64(params.ZeroBaseFee)
-
-	// b.SuggestPrice = unitPrice, for before Magma
-	//                = baseFee,   for after Magma
-	gasPrice, err := b.SuggestPrice(ctx)
-	if err != nil {
-		return err
-	}
-
-	// If user specifies both maxPriorityFee and maxFee, then we do not
-	// need to consult the chain for defaults. It's definitely a London tx.
-	if args.MaxPriorityFeePerGas == nil || args.MaxFeePerGas == nil {
-		if b.ChainConfig().IsEthTxTypeForkEnabled(head.Number) && args.GasPrice == nil {
-			if args.MaxPriorityFeePerGas == nil {
-				args.MaxPriorityFeePerGas = (*hexutil.Big)(gasPrice)
-			}
-			if args.MaxFeePerGas == nil {
-				// Before Magma hard fork, `gasFeeCap` was set to `baseFee*2 + maxPriorityFeePerGas` by default.
-				gasFeeCap := new(big.Int).Add(
-					(*big.Int)(args.MaxPriorityFeePerGas),
-					new(big.Int).Mul(fixedBaseFee, big.NewInt(2)),
-				)
-				if isMagma {
-					// After Magma hard fork, `gasFeeCap` was set to `baseFee*2` by default.
-					gasFeeCap = gasPrice
-				}
-				args.MaxFeePerGas = (*hexutil.Big)(gasFeeCap)
-			}
-			if isMagma {
-				if args.MaxFeePerGas.ToInt().Cmp(new(big.Int).Div(gasPrice, common.Big2)) < 0 {
-					return fmt.Errorf("maxFeePerGas (%v) < BaseFee (%v)", args.MaxFeePerGas, gasPrice)
-				}
-			} else if args.MaxPriorityFeePerGas.ToInt().Cmp(gasPrice) != 0 || args.MaxFeePerGas.ToInt().Cmp(gasPrice) != 0 {
-				return fmt.Errorf("only %s is allowed to be used as maxFeePerGas and maxPriorityPerGas", gasPrice.Text(16))
-			}
-			if args.MaxFeePerGas.ToInt().Cmp(args.MaxPriorityFeePerGas.ToInt()) < 0 {
-				return fmt.Errorf("maxFeePerGas (%v) < maxPriorityFeePerGas (%v)", args.MaxFeePerGas, args.MaxPriorityFeePerGas)
-			}
-		} else {
-			if args.MaxFeePerGas != nil || args.MaxPriorityFeePerGas != nil {
-				return errors.New("maxFeePerGas or maxPriorityFeePerGas specified but london is not active yet")
-			}
-			if args.GasPrice == nil {
-				// TODO-Klaytn: Original logic of Ethereum uses b.SuggestTipCap which suggests TipCap, not a GasPrice.
-				// But Klaytn currently uses fixed unit price determined by Governance, so using b.SuggestPrice
-				// is fine as now.
-				if b.ChainConfig().IsEthTxTypeForkEnabled(head.Number) {
-					// TODO-Klaytn: Klaytn is using fixed BaseFee(0) as now but
-					// if we apply dynamic BaseFee, we should add calculated BaseFee instead of params.ZeroBaseFee.
-					gasPrice.Add(gasPrice, new(big.Int).SetUint64(params.ZeroBaseFee))
-				}
-				args.GasPrice = (*hexutil.Big)(gasPrice)
-			}
-		}
-	} else {
-		// Both maxPriorityFee and maxFee set by caller. Sanity-check their internal relation
-		if isMagma {
-			if args.MaxFeePerGas.ToInt().Cmp(new(big.Int).Div(gasPrice, common.Big2)) < 0 {
-				return fmt.Errorf("maxFeePerGas (%v) < BaseFee (%v)", args.MaxFeePerGas, gasPrice)
-			}
-		} else {
-			if args.MaxFeePerGas.ToInt().Cmp(args.MaxPriorityFeePerGas.ToInt()) < 0 {
-				return fmt.Errorf("maxFeePerGas (%v) < maxPriorityFeePerGas (%v)", args.MaxFeePerGas, args.MaxPriorityFeePerGas)
-			}
-		}
-	}
-	if args.Value == nil {
-		args.Value = new(hexutil.Big)
-	}
-	if args.Nonce == nil {
-		nonce := b.GetPoolNonce(ctx, args.from())
-		args.Nonce = (*hexutil.Uint64)(&nonce)
-	}
-	if args.Data != nil && args.Input != nil && !bytes.Equal(*args.Data, *args.Input) {
-		return errors.New(`both "data" and "input" are set and not equal. Please use "input" to pass transaction call data`)
-	}
-	if args.To == nil && len(args.data()) == 0 {
-		return errors.New(`contract creation without any data provided`)
-	}
-	// Estimate the gas usage if necessary.
-	if args.Gas == nil {
-		// These fields are immutable during the estimation, safe to
-		// pass the pointer directly.
-		data := args.data()
-		callArgs := EthTransactionArgs{
-			From:                 args.From,
-			To:                   args.To,
-			GasPrice:             args.GasPrice,
-			MaxFeePerGas:         args.MaxFeePerGas,
-			MaxPriorityFeePerGas: args.MaxPriorityFeePerGas,
-			Value:                args.Value,
-			Data:                 (*hexutil.Bytes)(&data),
-			AccessList:           args.AccessList,
-		}
-		pendingBlockNr := rpc.NewBlockNumberOrHashWithNumber(rpc.PendingBlockNumber)
-		gasCap := uint64(0)
-		if rpcGasCap := b.RPCGasCap(); rpcGasCap != nil {
-			gasCap = rpcGasCap.Uint64()
-		}
-		estimated, err := EthDoEstimateGas(ctx, b, callArgs, pendingBlockNr, gasCap)
-		if err != nil {
-			return err
-		}
-		args.Gas = &estimated
-		logger.Trace("Estimate gas usage automatically", "gas", args.Gas)
-	}
-	if args.ChainID == nil {
-		id := (*hexutil.Big)(b.ChainConfig().ChainID)
-		args.ChainID = id
-	}
-	return nil
-}
-
-// ToMessage change EthTransactionArgs to types.Transaction in Klaytn.
-func (args *EthTransactionArgs) ToMessage(globalGasCap uint64, baseFee *big.Int, intrinsicGas uint64) (*types.Transaction, error) {
-	// Reject invalid combinations of pre- and post-1559 fee styles
-	if args.GasPrice != nil && (args.MaxFeePerGas != nil || args.MaxPriorityFeePerGas != nil) {
-		return nil, errors.New("both gasPrice and (maxFeePerGas or maxPriorityFeePerGas) specified")
-	} else if args.MaxFeePerGas != nil && args.MaxPriorityFeePerGas != nil {
-		if args.MaxFeePerGas.ToInt().Cmp(args.MaxPriorityFeePerGas.ToInt()) < 0 {
-			return nil, errors.New("MaxPriorityFeePerGas is greater than MaxFeePerGas")
-		}
-	}
-	// Set sender address or use zero address if none specified.
-	addr := args.from()
-
-	// Set default gas & gas price if none were set
-	gas := globalGasCap
-	if gas == 0 {
-		gas = uint64(math.MaxUint64 / 2)
-	}
-	if args.Gas != nil {
-		gas = uint64(*args.Gas)
-	}
-	if globalGasCap != 0 && globalGasCap < gas {
-		logger.Warn("Caller gas above allowance, capping", "requested", gas, "cap", globalGasCap)
-		gas = globalGasCap
-	}
-
-	// Do not update gasPrice unless any of args.GasPrice and args.MaxFeePerGas is specified.
-	gasPrice := new(big.Int)
-	if baseFee.Cmp(new(big.Int).SetUint64(params.ZeroBaseFee)) == 0 {
-		// If there's no basefee, then it must be a non-1559 execution
-		if args.GasPrice != nil {
-			gasPrice = args.GasPrice.ToInt()
-		} else if args.MaxFeePerGas != nil {
-			gasPrice = args.MaxFeePerGas.ToInt()
-		}
-	} else {
-		if args.GasPrice != nil {
-			gasPrice = args.GasPrice.ToInt()
-		} else if args.MaxFeePerGas != nil {
-			// User specified 1559 gas fields (or none), use those
-			gasPrice = args.MaxFeePerGas.ToInt()
-		} else {
-			// User specified neither GasPrice nor MaxFeePerGas, use baseFee
-			gasPrice = new(big.Int).Mul(baseFee, common.Big2)
-		}
-	}
-
-	value := new(big.Int)
-	if args.Value != nil {
-		value = args.Value.ToInt()
-	}
-	data := args.data()
-
-	// TODO-Klaytn: Klaytn does not support accessList yet.
-	// var accessList types.AccessList
-	// if args.AccessList != nil {
-	//	 accessList = *args.AccessList
-	// }
-	return types.NewMessage(addr, args.To, 0, value, gas, gasPrice, data, false, intrinsicGas), nil
-}
-
-// toTransaction converts the arguments to a transaction.
-// This assumes that setDefaults has been called.
-func (args *EthTransactionArgs) toTransaction() *types.Transaction {
-	var tx *types.Transaction
-	switch {
-	case args.MaxFeePerGas != nil:
-		al := types.AccessList{}
-		if args.AccessList != nil {
-			al = *args.AccessList
-		}
-		tx = types.NewTx(&types.TxInternalDataEthereumDynamicFee{
-			ChainID:      (*big.Int)(args.ChainID),
-			AccountNonce: uint64(*args.Nonce),
-			GasTipCap:    (*big.Int)(args.MaxPriorityFeePerGas),
-			GasFeeCap:    (*big.Int)(args.MaxFeePerGas),
-			GasLimit:     uint64(*args.Gas),
-			Recipient:    args.To,
-			Amount:       (*big.Int)(args.Value),
-			Payload:      args.data(),
-			AccessList:   al,
-		})
-	case args.AccessList != nil:
-		tx = types.NewTx(&types.TxInternalDataEthereumAccessList{
-			ChainID:      (*big.Int)(args.ChainID),
-			AccountNonce: uint64(*args.Nonce),
-			Recipient:    args.To,
-			GasLimit:     uint64(*args.Gas),
-			Price:        (*big.Int)(args.GasPrice),
-			Amount:       (*big.Int)(args.Value),
-			Payload:      args.data(),
-			AccessList:   *args.AccessList,
-		})
-	default:
-		tx = types.NewTx(&types.TxInternalDataLegacy{
-			AccountNonce: uint64(*args.Nonce),
-			Price:        (*big.Int)(args.GasPrice),
-			GasLimit:     uint64(*args.Gas),
-			Recipient:    args.To,
-			Amount:       (*big.Int)(args.Value),
-			Payload:      args.data(),
-		})
-	}
-	return tx
-}
-
-func (args *EthTransactionArgs) ToTransaction() *types.Transaction {
-	return args.toTransaction()
-}
-
 // SendTransaction creates a transaction for the given argument, sign it and submit it to the
 // transaction pool.
 func (api *EthereumAPI) SendTransaction(ctx context.Context, args EthTransactionArgs) (common.Hash, error) {
@@ -1319,7 +1041,7 @@ func (api *EthereumAPI) SendTransaction(ctx context.Context, args EthTransaction
 	if err := args.setDefaults(ctx, api.publicTransactionPoolAPI.b); err != nil {
 		return common.Hash{}, err
 	}
-	tx := args.toTransaction()
+	tx, _ := args.toTransaction()
 	signedTx, err := api.publicTransactionPoolAPI.sign(args.from(), tx)
 	if err != nil {
 		return common.Hash{}, err
@@ -1348,7 +1070,7 @@ func (api *EthereumAPI) FillTransaction(ctx context.Context, args EthTransaction
 		return nil, err
 	}
 	// Assemble the transaction and obtain rlp
-	tx := args.toTransaction()
+	tx, _ := args.toTransaction()
 	data, err := tx.MarshalBinary()
 	if err != nil {
 		return nil, err
@@ -1408,7 +1130,7 @@ func (api *EthereumAPI) SignTransaction(ctx context.Context, args EthTransaction
 		return nil, err
 	}
 	// Before actually sign the transaction, ensure the transaction fee is reasonable.
-	tx := args.toTransaction()
+	tx, _ := args.toTransaction()
 	if err := checkTxFee(tx.GasPrice(), tx.Gas(), b.RPCTxFeeCap()); err != nil {
 		return nil, err
 	}
@@ -1453,45 +1175,7 @@ func (api *EthereumAPI) PendingTransactions() ([]*EthRPCTransaction, error) {
 // Resend accepts an existing transaction and a new gas price and limit. It will remove
 // the given transaction from the pool and reinsert it with the new gas price and limit.
 func (api *EthereumAPI) Resend(ctx context.Context, sendArgs EthTransactionArgs, gasPrice *hexutil.Big, gasLimit *hexutil.Uint64) (common.Hash, error) {
-	if sendArgs.Nonce == nil {
-		return common.Hash{}, fmt.Errorf("missing transaction nonce in transaction spec")
-	}
-
-	if err := sendArgs.setDefaults(ctx, api.publicTransactionPoolAPI.b); err != nil {
-		return common.Hash{}, err
-	}
-
-	matchTx := sendArgs.toTransaction()
-	pending, err := api.publicTransactionPoolAPI.b.GetPoolTransactions()
-	if err != nil {
-		return common.Hash{}, err
-	}
-	for _, p := range pending {
-		signer := types.LatestSignerForChainID(p.ChainId())
-		wantSigHash := signer.Hash(matchTx)
-
-		if pFrom, err := types.Sender(signer, p); err == nil && pFrom == *sendArgs.From && signer.Hash(p) == wantSigHash {
-			// Match. Re-sign and send the transaction.
-			if gasPrice != nil && (*big.Int)(gasPrice).Sign() != 0 {
-				sendArgs.GasPrice = gasPrice
-			}
-			if gasLimit != nil && *gasLimit != 0 {
-				sendArgs.Gas = gasLimit
-			}
-			tx := sendArgs.toTransaction()
-
-			signedTx, err := api.publicTransactionPoolAPI.sign(*sendArgs.From, tx)
-			if err != nil {
-				return common.Hash{}, err
-			}
-			if err = api.publicTransactionPoolAPI.b.SendTx(ctx, signedTx); err != nil {
-				return common.Hash{}, err
-			}
-			return signedTx.Hash(), nil
-		}
-	}
-
-	return common.Hash{}, fmt.Errorf("Transaction %#x not found", matchTx.Hash())
+	return resend(api.publicTransactionPoolAPI, ctx, &sendArgs, gasPrice, gasLimit)
 }
 
 // Accounts returns the collection of accounts this node manages.
