+++ api/tx_args.go
@@ -35,6 +35,8 @@ import (
 	"github.com/klaytn/klaytn/blockchain/vm"
 	"github.com/klaytn/klaytn/common"
 	"github.com/klaytn/klaytn/common/hexutil"
+	"github.com/klaytn/klaytn/common/math"
+	"github.com/klaytn/klaytn/networks/rpc"
 	"github.com/klaytn/klaytn/params"
 	"github.com/klaytn/klaytn/rlp"
 )
@@ -105,6 +107,11 @@ type NewTxArgs interface {
 	setDefaults(context.Context, Backend) error
 	toTransaction() (*types.Transaction, error)
 	from() common.Address
+	gas() *hexutil.Uint64
+	gasPrice() *hexutil.Big
+	nonce() *hexutil.Uint64
+	setGas(*hexutil.Uint64)
+	setGasPrice(*hexutil.Big)
 }
 
 // SendTxArgs represents the arguments to submit a new transaction into the transaction pool.
@@ -360,6 +367,30 @@ func (args *SendTxArgs) toTransaction() (*types.Transaction, error) {
 	return types.NewTransactionWithMap(*args.TypeInt, values)
 }
 
+func (args *SendTxArgs) from() common.Address {
+	return args.From
+}
+
+func (args *SendTxArgs) gas() *hexutil.Uint64 {
+	return args.GasLimit
+}
+
+func (args *SendTxArgs) gasPrice() *hexutil.Big {
+	return args.Price
+}
+
+func (args *SendTxArgs) nonce() *hexutil.Uint64 {
+	return args.AccountNonce
+}
+
+func (args *SendTxArgs) setGas(gas *hexutil.Uint64) {
+	args.GasLimit = gas
+}
+
+func (args *SendTxArgs) setGasPrice(gasPrice *hexutil.Big) {
+	args.Price = gasPrice
+}
+
 type ValueTransferTxArgs struct {
 	From     common.Address  `json:"from"`
 	Gas      *hexutil.Uint64 `json:"gas"`
@@ -371,6 +402,26 @@ type ValueTransferTxArgs struct {
 
 func (args *ValueTransferTxArgs) from() common.Address {
 	return args.From
+}
+
+func (args *ValueTransferTxArgs) gas() *hexutil.Uint64 {
+	return args.Gas
+}
+
+func (args *ValueTransferTxArgs) gasPrice() *hexutil.Big {
+	return args.GasPrice
+}
+
+func (args *ValueTransferTxArgs) nonce() *hexutil.Uint64 {
+	return args.Nonce
+}
+
+func (args *ValueTransferTxArgs) setGas(gas *hexutil.Uint64) {
+	args.Gas = gas
+}
+
+func (args *ValueTransferTxArgs) setGasPrice(gasPrice *hexutil.Big) {
+	args.GasPrice = gasPrice
 }
 
 // setDefaults is a helper function that fills in default values for unspecified tx fields.
@@ -421,6 +472,26 @@ func (args *AccountUpdateTxArgs) from() common.Address {
 	return args.From
 }
 
+func (args *AccountUpdateTxArgs) gas() *hexutil.Uint64 {
+	return args.Gas
+}
+
+func (args *AccountUpdateTxArgs) gasPrice() *hexutil.Big {
+	return args.GasPrice
+}
+
+func (args *AccountUpdateTxArgs) nonce() *hexutil.Uint64 {
+	return args.Nonce
+}
+
+func (args *AccountUpdateTxArgs) setGas(gas *hexutil.Uint64) {
+	args.Gas = gas
+}
+
+func (args *AccountUpdateTxArgs) setGasPrice(gasPrice *hexutil.Big) {
+	args.GasPrice = gasPrice
+}
+
 // setDefaults is a helper function that fills in default values for unspecified tx fields.
 func (args *AccountUpdateTxArgs) setDefaults(ctx context.Context, b Backend) error {
 	if args.Gas == nil {
@@ -459,6 +530,302 @@ func (args *AccountUpdateTxArgs) toTransaction() (*types.Transaction, error) {
 	}
 
 	return tx, nil
+}
+
+// EthTransactionArgs represents the arguments to construct a new transaction
+// or a message call.
+// TransactionArgs in go-ethereum has been renamed to EthTransactionArgs.
+// TransactionArgs is defined in go-ethereum's internal package, so TransactionArgs is redefined here as EthTransactionArgs.
+type EthTransactionArgs struct {
+	From                 *common.Address `json:"from"`
+	To                   *common.Address `json:"to"`
+	Gas                  *hexutil.Uint64 `json:"gas"`
+	GasPrice             *hexutil.Big    `json:"gasPrice"`
+	MaxFeePerGas         *hexutil.Big    `json:"maxFeePerGas"`
+	MaxPriorityFeePerGas *hexutil.Big    `json:"maxPriorityFeePerGas"`
+	Value                *hexutil.Big    `json:"value"`
+	Nonce                *hexutil.Uint64 `json:"nonce"`
+
+	// We accept "data" and "input" for backwards-compatibility reasons.
+	// "input" is the newer name and should be preferred by clients.
+	// Issue detail: https://github.com/ethereum/go-ethereum/issues/15628
+	Data  *hexutil.Bytes `json:"data"`
+	Input *hexutil.Bytes `json:"input"`
+
+	// Introduced by AccessListTxType transaction.
+	AccessList *types.AccessList `json:"accessList,omitempty"`
+	ChainID    *hexutil.Big      `json:"chainId,omitempty"`
+}
+
+// from retrieves the transaction sender address.
+func (args *EthTransactionArgs) from() common.Address {
+	if args.From == nil {
+		return common.Address{}
+	}
+	return *args.From
+}
+
+func (args *EthTransactionArgs) gas() *hexutil.Uint64 {
+	return args.Gas
+}
+
+func (args *EthTransactionArgs) gasPrice() *hexutil.Big {
+	return args.GasPrice
+}
+
+func (args *EthTransactionArgs) nonce() *hexutil.Uint64 {
+	return args.Nonce
+}
+
+// data retrieves the transaction calldata. Input field is preferred.
+func (args *EthTransactionArgs) data() []byte {
+	if args.Input != nil {
+		return *args.Input
+	}
+	if args.Data != nil {
+		return *args.Data
+	}
+	return nil
+}
+
+func (args *EthTransactionArgs) setGas(gas *hexutil.Uint64) {
+	args.Gas = gas
+}
+
+func (args *EthTransactionArgs) setGasPrice(gasPrice *hexutil.Big) {
+	args.GasPrice = gasPrice
+}
+
+// setDefaults fills in default values for unspecified tx fields.
+func (args *EthTransactionArgs) setDefaults(ctx context.Context, b Backend) error {
+	if args.GasPrice != nil && (args.MaxFeePerGas != nil || args.MaxPriorityFeePerGas != nil) {
+		return errors.New("both gasPrice and (maxFeePerGas or maxPriorityFeePerGas) specified")
+	}
+	// After london, default to 1559 uncles gasPrice is set
+	head := b.CurrentBlock().Header()
+	isMagma := head.BaseFee != nil
+
+	fixedBaseFee := new(big.Int).SetUint64(params.ZeroBaseFee)
+
+	// b.SuggestPrice = unitPrice, for before Magma
+	//                = baseFee,   for after Magma
+	gasPrice, err := b.SuggestPrice(ctx)
+	if err != nil {
+		return err
+	}
+
+	// If user specifies both maxPriorityFee and maxFee, then we do not
+	// need to consult the chain for defaults. It's definitely a London tx.
+	if args.MaxPriorityFeePerGas == nil || args.MaxFeePerGas == nil {
+		if b.ChainConfig().IsEthTxTypeForkEnabled(head.Number) && args.GasPrice == nil {
+			if args.MaxPriorityFeePerGas == nil {
+				args.MaxPriorityFeePerGas = (*hexutil.Big)(gasPrice)
+			}
+			if args.MaxFeePerGas == nil {
+				// Before Magma hard fork, `gasFeeCap` was set to `baseFee*2 + maxPriorityFeePerGas` by default.
+				gasFeeCap := new(big.Int).Add(
+					(*big.Int)(args.MaxPriorityFeePerGas),
+					new(big.Int).Mul(fixedBaseFee, big.NewInt(2)),
+				)
+				if isMagma {
+					// After Magma hard fork, `gasFeeCap` was set to `baseFee*2` by default.
+					gasFeeCap = gasPrice
+				}
+				args.MaxFeePerGas = (*hexutil.Big)(gasFeeCap)
+			}
+			if isMagma {
+				if args.MaxFeePerGas.ToInt().Cmp(new(big.Int).Div(gasPrice, common.Big2)) < 0 {
+					return fmt.Errorf("maxFeePerGas (%v) < BaseFee (%v)", args.MaxFeePerGas, gasPrice)
+				}
+			} else if args.MaxPriorityFeePerGas.ToInt().Cmp(gasPrice) != 0 || args.MaxFeePerGas.ToInt().Cmp(gasPrice) != 0 {
+				return fmt.Errorf("only %s is allowed to be used as maxFeePerGas and maxPriorityPerGas", gasPrice.Text(16))
+			}
+			if args.MaxFeePerGas.ToInt().Cmp(args.MaxPriorityFeePerGas.ToInt()) < 0 {
+				return fmt.Errorf("maxFeePerGas (%v) < maxPriorityFeePerGas (%v)", args.MaxFeePerGas, args.MaxPriorityFeePerGas)
+			}
+		} else {
+			if args.MaxFeePerGas != nil || args.MaxPriorityFeePerGas != nil {
+				return errors.New("maxFeePerGas or maxPriorityFeePerGas specified but london is not active yet")
+			}
+			if args.GasPrice == nil {
+				// TODO-Klaytn: Original logic of Ethereum uses b.SuggestTipCap which suggests TipCap, not a GasPrice.
+				// But Klaytn currently uses fixed unit price determined by Governance, so using b.SuggestPrice
+				// is fine as now.
+				if b.ChainConfig().IsEthTxTypeForkEnabled(head.Number) {
+					// TODO-Klaytn: Klaytn is using fixed BaseFee(0) as now but
+					// if we apply dynamic BaseFee, we should add calculated BaseFee instead of params.ZeroBaseFee.
+					gasPrice.Add(gasPrice, new(big.Int).SetUint64(params.ZeroBaseFee))
+				}
+				args.GasPrice = (*hexutil.Big)(gasPrice)
+			}
+		}
+	} else {
+		// Both maxPriorityFee and maxFee set by caller. Sanity-check their internal relation
+		if isMagma {
+			if args.MaxFeePerGas.ToInt().Cmp(new(big.Int).Div(gasPrice, common.Big2)) < 0 {
+				return fmt.Errorf("maxFeePerGas (%v) < BaseFee (%v)", args.MaxFeePerGas, gasPrice)
+			}
+		} else {
+			if args.MaxFeePerGas.ToInt().Cmp(args.MaxPriorityFeePerGas.ToInt()) < 0 {
+				return fmt.Errorf("maxFeePerGas (%v) < maxPriorityFeePerGas (%v)", args.MaxFeePerGas, args.MaxPriorityFeePerGas)
+			}
+		}
+	}
+	if args.Value == nil {
+		args.Value = new(hexutil.Big)
+	}
+	if args.Nonce == nil {
+		nonce := b.GetPoolNonce(ctx, args.from())
+		args.Nonce = (*hexutil.Uint64)(&nonce)
+	}
+	if args.Data != nil && args.Input != nil && !bytes.Equal(*args.Data, *args.Input) {
+		return errors.New(`both "data" and "input" are set and not equal. Please use "input" to pass transaction call data`)
+	}
+	if args.To == nil && len(args.data()) == 0 {
+		return errors.New(`contract creation without any data provided`)
+	}
+	// Estimate the gas usage if necessary.
+	if args.Gas == nil {
+		// These fields are immutable during the estimation, safe to
+		// pass the pointer directly.
+		data := args.data()
+		callArgs := EthTransactionArgs{
+			From:                 args.From,
+			To:                   args.To,
+			GasPrice:             args.GasPrice,
+			MaxFeePerGas:         args.MaxFeePerGas,
+			MaxPriorityFeePerGas: args.MaxPriorityFeePerGas,
+			Value:                args.Value,
+			Data:                 (*hexutil.Bytes)(&data),
+			AccessList:           args.AccessList,
+		}
+		pendingBlockNr := rpc.NewBlockNumberOrHashWithNumber(rpc.PendingBlockNumber)
+		gasCap := uint64(0)
+		if rpcGasCap := b.RPCGasCap(); rpcGasCap != nil {
+			gasCap = rpcGasCap.Uint64()
+		}
+		estimated, err := EthDoEstimateGas(ctx, b, callArgs, pendingBlockNr, gasCap)
+		if err != nil {
+			return err
+		}
+		args.Gas = &estimated
+		logger.Trace("Estimate gas usage automatically", "gas", args.Gas)
+	}
+	if args.ChainID == nil {
+		id := (*hexutil.Big)(b.ChainConfig().ChainID)
+		args.ChainID = id
+	}
+	return nil
+}
+
+// ToMessage change EthTransactionArgs to types.Transaction in Klaytn.
+func (args *EthTransactionArgs) ToMessage(globalGasCap uint64, baseFee *big.Int, intrinsicGas uint64) (*types.Transaction, error) {
+	// Reject invalid combinations of pre- and post-1559 fee styles
+	if args.GasPrice != nil && (args.MaxFeePerGas != nil || args.MaxPriorityFeePerGas != nil) {
+		return nil, errors.New("both gasPrice and (maxFeePerGas or maxPriorityFeePerGas) specified")
+	} else if args.MaxFeePerGas != nil && args.MaxPriorityFeePerGas != nil {
+		if args.MaxFeePerGas.ToInt().Cmp(args.MaxPriorityFeePerGas.ToInt()) < 0 {
+			return nil, errors.New("MaxPriorityFeePerGas is greater than MaxFeePerGas")
+		}
+	}
+	// Set sender address or use zero address if none specified.
+	addr := args.from()
+
+	// Set default gas & gas price if none were set
+	gas := globalGasCap
+	if gas == 0 {
+		gas = uint64(math.MaxUint64 / 2)
+	}
+	if args.Gas != nil {
+		gas = uint64(*args.Gas)
+	}
+	if globalGasCap != 0 && globalGasCap < gas {
+		logger.Warn("Caller gas above allowance, capping", "requested", gas, "cap", globalGasCap)
+		gas = globalGasCap
+	}
+
+	// Do not update gasPrice unless any of args.GasPrice and args.MaxFeePerGas is specified.
+	gasPrice := new(big.Int)
+	if baseFee.Cmp(new(big.Int).SetUint64(params.ZeroBaseFee)) == 0 {
+		// If there's no basefee, then it must be a non-1559 execution
+		if args.GasPrice != nil {
+			gasPrice = args.GasPrice.ToInt()
+		} else if args.MaxFeePerGas != nil {
+			gasPrice = args.MaxFeePerGas.ToInt()
+		}
+	} else {
+		if args.GasPrice != nil {
+			gasPrice = args.GasPrice.ToInt()
+		} else if args.MaxFeePerGas != nil {
+			// User specified 1559 gas fields (or none), use those
+			gasPrice = args.MaxFeePerGas.ToInt()
+		} else {
+			// User specified neither GasPrice nor MaxFeePerGas, use baseFee
+			gasPrice = new(big.Int).Mul(baseFee, common.Big2)
+		}
+	}
+
+	value := new(big.Int)
+	if args.Value != nil {
+		value = args.Value.ToInt()
+	}
+	data := args.data()
+
+	// TODO-Klaytn: Klaytn does not support accessList yet.
+	// var accessList types.AccessList
+	// if args.AccessList != nil {
+	//	 accessList = *args.AccessList
+	// }
+	return types.NewMessage(addr, args.To, 0, value, gas, gasPrice, data, false, intrinsicGas), nil
+}
+
+// toTransaction converts the arguments to a transaction.
+// This assumes that setDefaults has been called.
+func (args *EthTransactionArgs) toTransaction() (*types.Transaction, error) {
+	var tx *types.Transaction
+	switch {
+	case args.MaxFeePerGas != nil:
+		al := types.AccessList{}
+		if args.AccessList != nil {
+			al = *args.AccessList
+		}
+		tx = types.NewTx(&types.TxInternalDataEthereumDynamicFee{
+			ChainID:      (*big.Int)(args.ChainID),
+			AccountNonce: uint64(*args.Nonce),
+			GasTipCap:    (*big.Int)(args.MaxPriorityFeePerGas),
+			GasFeeCap:    (*big.Int)(args.MaxFeePerGas),
+			GasLimit:     uint64(*args.Gas),
+			Recipient:    args.To,
+			Amount:       (*big.Int)(args.Value),
+			Payload:      args.data(),
+			AccessList:   al,
+		})
+	case args.AccessList != nil:
+		tx = types.NewTx(&types.TxInternalDataEthereumAccessList{
+			ChainID:      (*big.Int)(args.ChainID),
+			AccountNonce: uint64(*args.Nonce),
+			Recipient:    args.To,
+			GasLimit:     uint64(*args.Gas),
+			Price:        (*big.Int)(args.GasPrice),
+			Amount:       (*big.Int)(args.Value),
+			Payload:      args.data(),
+			AccessList:   *args.AccessList,
+		})
+	default:
+		tx = types.NewTx(&types.TxInternalDataLegacy{
+			AccountNonce: uint64(*args.Nonce),
+			Price:        (*big.Int)(args.GasPrice),
+			GasLimit:     uint64(*args.Gas),
+			Recipient:    args.To,
+			Amount:       (*big.Int)(args.Value),
+			Payload:      args.data(),
+		})
+	}
+	return tx, nil
+}
+
+func (args *EthTransactionArgs) ToTransaction() (*types.Transaction, error) {
+	return args.toTransaction()
 }
 
 // isReverted checks given error is vm.ErrExecutionReverted
